auth: auth directory structure/types
